refactor: move XREG wire constants to types.go
